# API_DOCUMENTATION.yaml
openapi: 3.0.3
info:
  title: FirstWeek - Complete API Documentation
  description: |
    # FirstWeek API

    Complete API documentation for the FirstWeek project including:
    - **Auth Service** (Port 3001) - User authentication and session management
    - **Chat Service** (Port 3002) - Real-time messaging and conversation management
    - **RAG API** (Port 8000) - Intelligent document retrieval and LLM responses

    ## Architecture

    ```
    Frontend → Auth Service (JWT) → Chat Service ⇄ RAG API
    ```

    ## Authentication

    Most endpoints require JWT authentication via cookies:
    - **Cookie**: `firstweek_avatar_access_token` (httpOnly)
    - **Header**: `Authorization: Bearer <token>` (alternative)

    ### Getting Started

    1. **Register**: `POST /api/users/register`
    2. **Login**: `POST /api/users/login` → Receives JWT tokens in cookies
    3. **Use Protected Endpoints**: Include cookies or Authorization header

    ## Service Ports

    - Auth Service: `http://localhost:3001`
    - Chat Service: `http://localhost:3002`
    - RAG API: `http://localhost:8000`

  version: 1.0.0
  contact:
    name: FIRSTWEEK Team
  license:
    name: ISC
servers:
  - url: http://localhost:3001/api/users
    description: Auth Service (Development)
  - url: http://localhost:3002/api/chat
    description: Chat Service (Development)
  - url: http://localhost:8000/api/v1
    description: RAG API (Development)
  - url: https://api.example.invalid
    description: Production
...
paths:
...
  /register:
    post:
      tags: [Authentication]
      summary: Register new user
      description: |
        Create a new user account with role assignment.

        **Default Role**: EMPLOYEE
        **Available Roles**: GUEST, EMPLOYEE, EXECUTIVE, COMPANY_ADMIN, SUPER_ADMIN

        **Returns**: User profile + JWT tokens in httpOnly cookies
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, username, firstName, lastName]
              properties:
                email:
                  type: string
                  format: email
                  example: employee@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  description: Must contain uppercase, lowercase, number, and special character
                  example: SecurePass123!
                username:
                  type: string
                  minLength: 3
                  maxLength: 30
                  pattern: ^[a-zA-Z0-9_]+$
                  example: john_doe
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: Doe
                role:
                  type: string
                  enum: [GUEST, EMPLOYEE, EXECUTIVE, COMPANY_ADMIN, SUPER_ADMIN]
                  default: EMPLOYEE
                  example: EMPLOYEE
                profilePic:
                  type: string
                  format: uri
                  nullable: true
      responses:
        '201':
          description: User registered successfully
          headers:
            Set-Cookie:
              description: JWT tokens (access + refresh)
              schema:
                type: string
                example: firstweek_avatar_access_token=eyJ...; HttpOnly; Secure; SameSite=Strict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
  /login:
    post:
      tags: [Authentication]
      summary: Login user
      description: |
        Authenticate user with email/username and password.

        **Returns**: User profile + JWT tokens in httpOnly cookies
        **Token Lifespan**:
        - Access token: 15 minutes
        - Refresh token: 7 days
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier:
                  type: string
                  description: Email or username
                  example: john_doe
                password:
                  type: string
                  format: password
                  example: SecurePass123!
                rememberMe:
                  type: boolean
                  default: false
                  description: Extend refresh token to 30 days
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
  /logout:
    post:
      tags: [Authentication]
      summary: Logout user
      description: Invalidate current session and clear cookies
      operationId: logoutUser
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Logout successful
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Get new access token using refresh token from cookie
      operationId: refreshToken
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  description: Refresh token (optional if in cookie)
      responses:
        '200':
          description: Token refreshed successfully
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  accessToken:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /me:
    get:
      tags: [User Profile]
      summary: Get current user profile
      description: Retrieve authenticated user's profile information
      operationId: getCurrentUser
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    put:
      tags: [User Profile]
      summary: Update user profile
      description: Update current user's profile information
      operationId: updateUserProfile
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                profilePic:
                  type: string
                  format: uri
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /change-password:
    put:
      tags: [Password Management]
      summary: Change password
      description: Change password for authenticated user
      operationId: changePassword
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                  format: password
                newPassword:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '200':
          description: Password changed successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /password/forgot:
    post:
      tags: [Password Management]
      summary: Request password reset
      description: Send password reset link to user's email
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Reset email sent (always returns success for security)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: If the email exists, a reset link has been sent
  /password/reset:
    post:
      tags: [Password Management]
      summary: Reset password
      description: Reset password using token from email
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token:
                  type: string
                  description: Reset token from email
                newPassword:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '200':
          description: Password reset successful
        '400':
          description: Invalid or expired token
  /sessions:
    get:
      tags: [Sessions]
      summary: Get active sessions
      description: List all active sessions for current user
      operationId: getUserSessions
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: Sessions retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      sessions:
                        type: array
                        items:
                          $ref: '#/components/schemas/Session'
  /sessions/{sessionId}:
    delete:
      tags: [Sessions]
      summary: Revoke session
      description: Revoke a specific session (logout from that device)
      operationId: revokeSession
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session revoked successfully
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ============================================================
  # CHAT SERVICE ENDPOINTS
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token in Authorization header

    cookieAuth:
      type: apiKey
      in: cookie
      name: firstweek_avatar_access_token
      description: JWT access token in httpOnly cookie
...
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: clx123abc456
        email:
          type: string
          format: email
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        profilePic:
          type: string
          format: uri
          nullable: true
        role:
          type: string
          enum: [GUEST, EMPLOYEE, EXECUTIVE, COMPANY_ADMIN, SUPER_ADMIN]
        emailVerified:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
    Session:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        deviceInfo:
          type: string
        ipAddress:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        lastActivityAt:
          type: string
          format: date-time
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              message: Validation failed
              code: VALIDATION_ERROR
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              message: Authentication required
              code: UNAUTHORIZED
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              message: Resource not found
              code: NOT_FOUND
    ConflictError:
      description: Resource conflict (duplicate)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              message: Email already exists
              code: CONFLICT
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              message: Internal server error
              code: INTERNAL_ERROR